% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisGetOutputRequestTypes}
\alias{CapsisGetOutputRequestTypes}
\title{Provide the Different Outputs of Capsis Variant}
\usage{
CapsisGetOutputRequestTypes(variant)
}
\arguments{
\item{variant}{a character string}
}
\value{
a vector with the different outputs
}
\description{
Provide the Different Outputs of Capsis Variant
}
\seealso{
[CapsisGetVariantList()] for the list of available variants
}
