% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisGetVariantFields}
\alias{CapsisGetVariantFields}
\title{Provide a List of Fields Required by a Variant}
\usage{
CapsisGetVariantFields(variant)
}
\arguments{
\item{variant}{a valid variant (string)}
}
\value{
a list of data.frame instances containing the information on
the fields
}
\description{
Provide a List of Fields Required by a Variant
}
