% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisGetVariantScope}
\alias{CapsisGetVariantScope}
\title{Provide the Scope of a Variant}
\usage{
CapsisGetVariantScope(variant)
}
\arguments{
\item{variant}{a valid variant (string)}
}
\value{
a list of variables with their scope
}
\description{
Provide the Scope of a Variant
}
