% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisGetVariantSpecies}
\alias{CapsisGetVariantSpecies}
\title{Provide the species for}
\usage{
CapsisGetVariantSpecies(variant, type = "All")
}
\arguments{
\item{variant}{a valid variant (string)}

\item{type}{either All, Broadleaved, or Coniferous}
}
\value{
a vector of characters
}
\description{
Provide the species for
}
