% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisSimulate}
\alias{CapsisSimulate}
\title{Run a Simulation with Capsis}
\usage{
CapsisSimulate(
  data,
  outputRequestList,
  variant,
  years,
  initialYear,
  isStochastic,
  nbRealizations,
  climateChange,
  applicationScale
)
}
\arguments{
\item{data}{a string in CSV format that represents the input data to run the
simulation on. The colnames should match the required fieldnames.}

\item{outputRequestList}{an object of type OutputRequestList that contains
the output data requested}

\item{variant}{a string containing the variant name to use for simulation}

\item{years}{an int containing the number of years the simulation should use}

\item{initialYear}{the initial date of the simulation}

\item{isStochastic}{a logical (true to enable stochastic simulation)}

\item{nbRealizations}{an integer standing for the number of realizations in
case of stochastic simulation [1-500]}

\item{climateChange}{astring indicating the climate change scenario. One
among these: NO_CHANGE, RCP2_6, RCP4_5, RCP6_0, or RCP8_5}

\item{applicationScale}{a string indicating the scale of the simulation.
Either Stand or FMU.}
}
\value{
a SimulationResult instance
}
\description{
Run a Simulation with Capsis
}
