% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\name{CapsisStatus}
\alias{CapsisStatus}
\title{Provide the Status of OSM Web API}
\usage{
CapsisStatus()
}
\value{
a list of data.frame instances containing the information on
the fields
}
\description{
Provide the Status of OSM Web API
}
