% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capsis.R
\docType{data}
\name{STR_RE2_70}
\alias{STR_RE2_70}
\title{A list of plots that can be passed to Capsis Web API}
\format{
An object of class \code{data.frame} with 2047 rows and 16 columns.
}
\usage{
data(STR_RE2_70)
}
\description{
A list of plots that can be passed to Capsis Web API
}
\examples{
data(STR_RE2_70)
}
\keyword{datasets}
